requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pyyaml==6.0.1
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langsmith>=0.2.0
langgraph>=0.2.0
openai>=1.6.1
notion-client==2.2.1
pydantic>=2.7.0
pydantic-settings>=2.1.0
python-multipart==0.0.6
httpx==0.25.2
aiofiles==23.2.1
pillow==10.1.0
jinja2==3.1.2
chromadb==0.4.22
beautifulsoup4==4.12.3
lxml==5.3.0
apscheduler==3.10.4
websockets==12.0

